+++ test/core-filters/filterFiles-skipIfNotExists.test.js
@@ -1,0 +1,41 @@
+const corefilters = require("../../lib/core-filters");
+const fs1 = require('../../lib/core-util');
+const mockdata = require('./mock-data')
+
+test("filter/setupFilters/skipIfNotExists/fileExist", () => {
+    var workingObject = {
+        validatedOption: { localPath: '.' },
+        fileFilters: [corefilters.skipIfNotExists],
+        remoteFileList: mockdata.mockRemoteFileList,
+        filteredFileList: []
+    };
+    jest.spyOn(fs1, "isPathExist").mockReturnValue(true);
+
+    return expect(corefilters.filterFiles(workingObject))
+        .resolves
+        .toMatchObject({
+            filteredFileList: [
+                { name: 'Mock_File_1.zip' },
+                { name: 'Mock_File_2.zip' }
+            ]
+        });
+});
+
+test("filter/setupFilters/skipIfNotExists/fileNotExist", () => {
+    var workingObject = {
+        validatedOption: { localPath: '.' },
+        fileFilters: [corefilters.skipIfNotExists],
+        remoteFileList: mockdata.mockRemoteFileList,
+        filteredFileList: []
+    };
+    jest.spyOn(fs1, "isPathExist").mockReturnValue(false);
+
+    return expect(corefilters.filterFiles(workingObject))
+        .resolves.not
+        .toMatchObject({
+            filteredFileList: [
+                { name: 'Mock_File_1.zip' },
+                { name: 'Mock_File_2.zip' }
+            ]
+        });
+});